// interfaces/WeatherForecastData.interface.ts
import { TWeatherStatus } from './WeatherData.interface'

export interface IWeatherForecast_ListItemData {
  dt: number
  main: {
    temp: number
    feels_like: number
    temp_min: number
    temp_max: number
    pressure: number
    sea_level: number
    grnd_level: number
    humidity: number
    temp_kf: number
  }
  weather: {
    id: number
    main: TWeatherStatus
    description: string
    icon: string
  }[]
  clouds: {
    all: number
  }
  wind: {
    speed: number
    deg: number
    gust: number
  }

  visibility: number
  pop: number
  sys: {
    [key: string]: string
  }
  dt_txt: Date
  rain?: {
    [time: string]: number
  }
}

export interface IWeatherForecastData {
  cod: string
  message: number
  cnt: number
  list: IWeatherForecast_ListItemData[]
  city: {
    id: number
    name: string
    coord: {
      lat: number
      lon: number
    }
    country: string
    population: number
    timezone: number
    sunrise: number
    sunset: number
  }
}